+++ backend/systeme-reservation-jo/src/main/java/com/example/systeme_reservation_jo/repository/UtilisateurRepository.java
@@ -4,7 +4,7 @@ import com.example.systeme_reservation_jo.model.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UtilisateurRepository extends JpaRepository<Utilisateur, Long> {
+public interface UtilisateurRepository extends JpaRepository<Utilisateur, Integer> {
     Optional<Utilisateur> findByEmail(String email); // Recherche par email
     Boolean existsByEmail(String email);          // Vérifie si un email existe
     Optional<Utilisateur> findByUsername(String username); //On garde la méthode au cas où.
